requirements.txt
fastapi
uvicorn[standard]
openai
langchain
psycopg2-binary
pgvector
python-dotenv
structlog
fdb
